pyproject.toml
[project]
name = "memoreason"
version = "0.1.0"
authors = [{ name = "Anonymous authors" }]
description = "Dataset construction, fictionalization, and evaluation code for the MemoReason benchmark."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "PyYAML>=6.0",
    "pydantic>=2.8",
    "num2words>=0.5.14",
    "numpy>=1.26",
    "pandas>=2.0",
    "matplotlib>=3.8",
    "statsmodels>=0.14",
    "scipy>=1.11",
    "tqdm>=4.66",
    "datasets>=2.19",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.40",
    "requests>=2.32",
    "transformers>=4.45",
    "torch>=2.2",
]
web = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "jinja2>=3.1,<4.0",
    "bcrypt>=4.1,<5.0",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src", "src.*", "web", "web.*"]
